+++ tests/utils/transfer.py
@@ -1,5 +1,6 @@
 #coding: utf-8
 from copy import deepcopy
+from datetime import date, timedelta
 from random import randint
 from starkbank import Transfer
 from .names.names import get_full_name
@@ -16,7 +17,7 @@ example_transfer = Transfer(
 )
 
 
-def generateExampleTransfersJson(n=1):
+def generateExampleTransfersJson(n=1, randomSchedule=False):
     transfers = []
     for _ in range(n):
         amount = randint(1, 10)
@@ -24,5 +25,7 @@ def generateExampleTransfersJson(n=1):
         transfer.name = get_full_name()
         transfer.amount = amount
         transfer.tax_id = generateCpf() if randint(0, 1) else generateCnpj()
+        if randomSchedule:
+            transfer.scheduled = date.today() + timedelta(days=randint(0, 10))
         transfers.append(transfer)
     return transfers